Broke subpath impl across multiple files. More sylistic changes per review
