+++ bezier-rs/lib/src/subpath/core.rs
@@ -1,101 +1,4 @@
-use std::ops::{Index, IndexMut};
-
-use glam::DVec2;
-
-use crate::Bezier;
-
-pub struct ManipulatorGroup {
-	pub anchor: DVec2,
-	pub in_handle: Option<DVec2>,
-	pub out_handle: Option<DVec2>,
-}
-
-pub struct SubPath {
-	manipulator_groups: Vec<ManipulatorGroup>,
-	closed: bool,
-}
-
-impl Index<usize> for SubPath {
-	type Output = ManipulatorGroup;
-
-	fn index(&self, index: usize) -> &Self::Output {
-		assert!(index < self.len());
-		&self.manipulator_groups[index]
-	}
-}
-
-impl IndexMut<usize> for SubPath {
-	fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-		assert!(index < self.len());
-		&mut self.manipulator_groups[index]
-	}
-}
-
-pub struct SubPathIter<'a> {
-	index: usize,
-	sub_path: &'a SubPath,
-}
-
-impl Iterator for SubPathIter<'_> {
-	type Item = Bezier;
-
-	fn next(&mut self) -> Option<Self::Item> {
-		if self.index >= self.sub_path.len() - 1 + (self.sub_path.closed as usize) {
-			return None;
-		}
-		let start_index = self.index;
-		let end_index = (self.index + 1) % self.sub_path.len();
-		self.index += 1;
-
-		let start = self.sub_path[start_index].anchor;
-		let end = self.sub_path[end_index].anchor;
-		let handle1 = self.sub_path[start_index].out_handle;
-		let handle2 = self.sub_path[end_index].in_handle;
-
-		if handle1.is_none() && handle2.is_none() {
-			return Some(Bezier::from_linear_dvec2(start, end));
-		}
-		if handle1.is_none() || handle2.is_none() {
-			return Some(Bezier::from_quadratic_dvec2(start, handle1.or(handle2).unwrap(), end));
-		}
-		Some(Bezier::from_cubic_dvec2(start, handle1.unwrap(), handle2.unwrap(), end))
-	}
-}
-
-/// Struct to represent optional parameters that can be passed to the `into_svg` function.
-pub struct ToSVGOptions {
-	pub curve_stroke_color: String,
-	pub curve_stroke_width: f64,
-	pub anchor_stroke_color: String,
-	pub anchor_stroke_width: f64,
-	pub anchor_radius: f64,
-	pub anchor_fill: String,
-	pub handle_line_stroke_color: String,
-	pub handle_line_stroke_width: f64,
-	pub handle_point_stroke: String,
-	pub handle_point_radius: f64,
-	pub handle_point_stroke_width: f64,
-	pub handle_point_fill: String,
-}
-
-impl Default for ToSVGOptions {
-	fn default() -> Self {
-		ToSVGOptions {
-			curve_stroke_color: String::from("black"),
-			curve_stroke_width: 2.,
-			anchor_stroke_color: String::from("black"),
-			anchor_stroke_width: 2.,
-			anchor_radius: 4.,
-			anchor_fill: String::from("white"),
-			handle_line_stroke_color: String::from("grey"),
-			handle_line_stroke_width: 1.,
-			handle_point_stroke: String::from("grey"),
-			handle_point_stroke_width: 1.5,
-			handle_point_radius: 3.,
-			handle_point_fill: String::from("white"),
-		}
-	}
-}
+use super::*;
 
 impl SubPath {
 	/// Create a new SubPath using a list of ManipulatorGroups.
@@ -139,6 +42,7 @@ impl SubPath {
 		SubPathIter { sub_path: self, index: 0 }
 	}
 
+	/// Returns an SVG representation of the `SubPath`.
 	pub fn to_svg(&self, options: ToSVGOptions) -> String {
 		if self.is_empty() {
 			return String::new();
@@ -151,7 +55,7 @@ impl SubPath {
 		);
 		let handle_point_options = format!(
 			r#"r="{}", stroke="{}" stroke-width="{}" fill="{}""#,
-			options.handle_point_radius, options.handle_point_stroke, options.handle_point_stroke_width, options.handle_point_fill
+			options.handle_point_radius, options.handle_point_stroke_color, options.handle_point_stroke_width, options.handle_point_fill
 		);
 		let handle_line_options = format!(
 			r#"stroke="{}" stroke-width="{}" fill="transparent""#,
@@ -175,29 +79,25 @@ impl SubPath {
 			let handle1 = self[start_index].out_handle;
 			let handle2 = self[end_index].in_handle;
 
-			if handle1.is_some() && handle2.is_some() {
-				handle_pieces.push(format!("M {} {} L {} {}", start.x, start.y, handle1.unwrap().x, handle1.unwrap().y));
-				handle_circles.push(format!(r#"<circle cx="{}" cy="{}" {}/>"#, handle1.unwrap().x, handle1.unwrap().y, handle_point_options));
-				handle_pieces.push(format!("M {} {} L {} {}", end.x, end.y, handle2.unwrap().x, handle2.unwrap().y));
-				handle_circles.push(format!(r#"<circle cx="{}" cy="{}" {}/>"#, handle2.unwrap().x, handle2.unwrap().y, handle_point_options));
-			} else if handle1.is_some() || handle2.is_some() {
-				let quad_handle = handle1.or(handle2).unwrap();
-				handle_pieces.push(format!("M {} {} L {} {}", start.x, start.y, quad_handle.x, quad_handle.y));
-				handle_pieces.push(format!("M {} {} L {} {}", end.x, end.y, quad_handle.x, quad_handle.y));
-				handle_circles.push(format!(r#"<circle cx="{}" cy="{}" {}/>"#, quad_handle.x, quad_handle.y, handle_point_options));
+			let main_path: String;
+			if let (Some(h1), Some(h2)) = (handle1, handle2) {
+				handle_pieces.push(format!("M {} {} L {} {}", start.x, start.y, h1.x, h1.y));
+				handle_circles.push(format!(r#"<circle cx="{}" cy="{}" {}/>"#, h1.x, h1.y, handle_point_options));
+				handle_pieces.push(format!("M {} {} L {} {}", end.x, end.y, h2.x, h2.y));
+				handle_circles.push(format!(r#"<circle cx="{}" cy="{}" {}/>"#, h2.x, h2.y, handle_point_options));
+				main_path = format!("C {} {} {} {}", h1.x, h1.y, h2.x, h2.y);
+			} else if let Some(handle) = handle1.or(handle2) {
+				handle_pieces.push(format!("M {} {} L {} {}", start.x, start.y, handle.x, handle.y));
+				handle_pieces.push(format!("M {} {} L {} {}", end.x, end.y, handle.x, handle.y));
+				handle_circles.push(format!(r#"<circle cx="{}" cy="{}" {}/>"#, handle.x, handle.y, handle_point_options));
+				main_path = format!("Q {} {}", handle.x, handle.y);
+			} else {
+				main_path = String::from("L");
 			}
-
-			let main_path = {
-				if handle1.is_none() && handle2.is_none() {
-					String::from("L")
-				} else if handle1.is_none() || handle2.is_none() {
-					let handle = handle1.or(handle2).unwrap();
-					format!("Q {} {}", handle.x, handle.y)
-				} else {
-					format!("C {} {} {} {}", handle1.unwrap().x, handle1.unwrap().y, handle2.unwrap().x, handle2.unwrap().y)
-				}
-			};
 			path_pieces.push(format!("{} {} {}", main_path, end.x, end.y));
+		}
+		if self.closed {
+			path_pieces.push(String::from("Z"));
 		}
 
 		format!(
@@ -210,22 +110,16 @@ impl SubPath {
 			anchor_circles.join(""),
 		)
 	}
-
-	/// Return the sum of the approximation of the length of each bezier curve along the subpath.
-	/// - `num_subdivisions` - Number of subdivisions used to approximate the curve. The default value is 1000.
-	pub fn length(&self, num_subdivisions: Option<i32>) -> f64 {
-		self.iter().map(|bezier| bezier.length(num_subdivisions)).sum()
-	}
 }
 
 #[cfg(test)]
 mod tests {
 
-	use glam::DVec2;
-
 	use crate::Bezier;
 
 	use super::*;
+
+	use glam::DVec2;
 
 	#[test]
 	fn length_quadratic() {
@@ -301,11 +195,9 @@ mod tests {
 			],
 			false,
 		);
-
 		assert_eq!(subpath.length(None), bezier1.length(None) + bezier2.length(None));
 
 		subpath.closed = true;
-
 		assert_eq!(subpath.length(None), bezier1.length(None) + bezier2.length(None) + bezier3.length(None));
 	}
 }
